Chad Mode + and More! Baths should look better now. Arrows shafts now are produced 3x. Arrowheads now produce 4x. Mining suprise fixed. Chad Mode + is added to the gamemode vote. bugfixes and typo fixes.  Changes to be committed: 	modified:   code/datums/vote.dm 	modified:   code/game/objects/items/stacks/stack.dm 	modified:   code/game/objects/map_metadata/_map_metadata.dm 	modified:   code/game/turfs/floor_attackby.dm 	modified:   code/modules/1713/researchcost.dm 	modified:   code/modules/materials/materials.dm 	modified:   code/modules/mob/living/carbon/human/human_defines.dm 	modified:   code/modules/mob/living/carbon/human/human_life.dm 	modified:   code/processes/mapswap.dm 	modified:   config/material_recipes.txt 	modified:   icons/mob/human.dmi 	modified:   icons/mob/human_face.dmi 	modified:   icons/mob/human_races/r_human.dmi 	modified:   icons/obj/bathtub.dmi 	modified:   icons/obj/bathtub_big.dmi 	modified:   icons/obj/structures.dmi

+++ code/modules/1713/researchcost.dm
@@ -14,6 +14,11 @@
 	not_movable = FALSE
 	not_disassemblable = TRUE
 
+/obj/structure/researchdesk/New()
+	..()
+	if(map.chad_mode_plus)
+		icon_state = "altar_of_chad"
+
 /obj/structure/researchdesk/attackby(var/obj/item/W as obj, var/mob/living/carbon/human/H as mob)
 	if (!map.civilizations || map.ID == MAP_CIVILIZATIONS || map.ID == MAP_TRIBES)
 		return
@@ -25,17 +30,30 @@
 		H << "You are not part of any factions."
 		return
 
-	if (W.value == 0)
-		H << "This item has no research value."
-		return
+	if(!map.chad_mode_plus) //if not chad mode, normal value check
+		if (W.value == 0)
+			H << "This item has no research value."
+			return
+	else //if is chad mode, only humans will suffice.
+		if (!istype(W, /mob/living/carbon/human))
+			H << "This item has no research value."
+			return
 
-	if  (done == FALSE)
-		if (istype(W, /obj/item/stack))
-			marketval = W.value
-			moneyin = (((marketval*W.amount)/100)*2)
-		else
-			marketval = W.value
-			moneyin = ((marketval/100)*2)
+	if (done == FALSE)
+		if (!map.chad_mode_plus) //Normal research
+			if (istype(W, /obj/item/stack))
+				marketval = W.value
+				moneyin = (((marketval*W.amount)/100)*2)
+			else
+				marketval = W.value
+				moneyin = ((marketval/100)*2)
+		else //If it IS Chad mode plus.
+			if (istype(W, /mob/living/carbon/human))
+				marketval = 8
+				moneyin = 8
+			else
+				H << "<span class='alert'>Only humans will suffice.</span>"
+				return
 		done = TRUE
 		var/list/display = list("Industrial", "Military", "Health", "Cancel")
 		var/choice = WWinput(H, "This is worth [moneyin] research points. Which category to research?", "Research Desk", "Cancel", display)
@@ -44,17 +62,32 @@
 			return
 		else if (choice == "Industrial" && W)
 			map.custom_civs[H.civilization][1] += (moneyin)
-			qdel(W)
+			if(map.chad_mode_plus)
+				var/mob/living/carbon/human/SACRIFICE as mob
+				SACRIFICE = W
+				SACRIFICE.gib()
+			else
+				qdel(W)
 			done = FALSE
 			return
 		else if (choice == "Military" && W)
 			map.custom_civs[H.civilization][2] += (moneyin)
-			qdel(W)
+			if(map.chad_mode_plus)
+				var/mob/living/carbon/human/SACRIFICE as mob
+				SACRIFICE = W
+				SACRIFICE.gib()
+			else
+				qdel(W)
 			done = FALSE
 			return
 		else if (choice == "Health" && W)
 			map.custom_civs[H.civilization][3] += (moneyin)
-			qdel(W)
+			if(map.chad_mode_plus)
+				var/mob/living/carbon/human/SACRIFICE as mob
+				SACRIFICE = W
+				SACRIFICE.gib()
+			else
+				qdel(W)
 			done = FALSE
 			return
 		else
